Make various zip/APK symbolization tests end-to-end tests
As motivated in commit d8c994d5c08e ("Make BPF symbolization tests
end-to-end tests"), in order to stand a chance of having a reasonable
grasp on what part of the public API surface is being tested, it makes
the most sense to make all end-to-end tests co-located next to each
other (and named consistently etc.).
With this change we take another step into this direction and move
zip/APK related symbolization tests into our end-to-end test suite.
